refactor: move viewport to new file
